+++ crates/rattler_cache/src/package_cache/cache_key.rs
@@ -1,6 +1,6 @@
 use rattler_conda_types::package::ArchiveIdentifier;
 use rattler_conda_types::PackageRecord;
-use rattler_digest::{compute_bytes_digest, compute_url_digest, Sha256, Sha256Hash};
+use rattler_digest::{compute_bytes_digest, compute_url_digest, Md5Hash, Sha256, Sha256Hash};
 use std::{
     fmt::{Display, Formatter},
     path::Path,
@@ -14,6 +14,7 @@ pub struct CacheKey {
     pub(crate) version: String,
     pub(crate) build_string: String,
     pub(crate) sha256: Option<Sha256Hash>,
+    pub(crate) md5: Option<Md5Hash>,
     pub(crate) origin_hash: Option<String>,
 }
 
@@ -27,6 +28,18 @@ impl CacheKey {
     /// Potentially adds a sha256 hash of the archive.
     pub fn with_opt_sha256(mut self, sha256: Option<Sha256Hash>) -> Self {
         self.sha256 = sha256;
+        self
+    }
+
+    /// Adds a md5 hash of the archive.
+    pub fn with_md5(mut self, md5: Md5Hash) -> Self {
+        self.md5 = Some(md5);
+        self
+    }
+
+    /// Potentially adds a md5 hash of the archive.
+    pub fn with_opt_md5(mut self, md5: Option<Md5Hash>) -> Self {
+        self.md5 = md5;
         self
     }
 
@@ -50,6 +63,11 @@ impl CacheKey {
     pub fn sha256(&self) -> Option<Sha256Hash> {
         self.sha256
     }
+
+    /// Return the md5 hash of the package if it is known.
+    pub fn md5(&self) -> Option<Md5Hash> {
+        self.md5
+    }
 }
 
 impl From<ArchiveIdentifier> for CacheKey {
@@ -59,6 +77,7 @@ impl From<ArchiveIdentifier> for CacheKey {
             version: pkg.version,
             build_string: pkg.build_string,
             sha256: None,
+            md5: None,
             origin_hash: None,
         }
     }
@@ -71,6 +90,7 @@ impl From<&PackageRecord> for CacheKey {
             version: record.version.to_string(),
             build_string: record.build.clone(),
             sha256: record.sha256,
+            md5: record.md5,
             origin_hash: None,
         }
     }