+++ py-rattler/rattler/__init__.py
@@ -22,9 +22,10 @@ from rattler.package import (
     PrefixPlaceholder,
     FileMode,
     IndexJson,
+    NoArchType,
+    NoArchLiteral,
 )
-from rattler.package.no_arch_type import NoArchType
-from rattler.prefix import PrefixRecord, PrefixPaths, PrefixPathsEntry, PrefixPathType
+from rattler.prefix import PrefixRecord, PrefixPaths, PrefixPathsEntry, PrefixPathType, Link, LinkType
 from rattler.platform import Platform
 from rattler.utils.rattler_version import get_rattler_version as _get_rattler_version
 from rattler.install import install
@@ -94,4 +95,7 @@ __all__ = [
     "Gateway",
     "SourceConfig",
     "NoArchType",
+    "NoArchLiteral",
+    "Link",
+    "LinkType",
 ]